+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   `@lumx/react`:
     -   `List`: fix jsdoc for deprecated `onListItemSelected` prop (removed in v4.0.0)
     -   `List`: remove default value for `tabIndex`
+    -   `ListItem`: alias `onItemSelected` prop to `onClick`
 
 ### Added